Moved DeliveryFormsAggregationHelperto own file
