+++ custom/icds_reports/models/aggregate.py
@@ -15,7 +15,6 @@ from custom.icds_reports.utils.aggregation import (
     AwcInfrastructureAggregationHelper,
     CcsRecordMonthlyAggregationHelper, ChildHealthMonthlyAggregationHelper,
     DailyFeedingFormsChildHealthAggregationHelper,
-    DeliveryFormsAggregationHelper,
     InactiveAwwsAggregationHelper,
     AwwIncentiveAggregationHelper
 )
@@ -28,6 +27,7 @@ from custom.icds_reports.utils.aggregation_helpers.birth_preparedness_forms impo
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms import ComplementaryFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms_ccs_record import \
     ComplementaryFormsCcsRecordAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.delivery_forms import DeliveryFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.growth_monitoring_forms import GrowthMonitoringFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.postnatal_care_forms_ccs_record import \
     PostnatalCareFormsCcsRecordAggregationHelper
